update insert api

+++ src/main/java/application/repository/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface UserRepository extends MongoRepository<User, String>, UserRepositoryCustom {
 
-    long countByEmail(String email);
+    int countByEmail(String email);
 }
